backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
earthengine-api==0.1.384
gunicorn==21.2.0
python-dotenv
